requirements.txt
discord.py==2.6.3
pyTelegramBotAPI
asyncpg
python-dotenv